% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rarestr.R
\name{plot.rarestr}
\alias{plot.rarestr}
\title{Plot fitted curve for TES and TESS}
\usage{
\method{plot}{rarestr}(x)
}
\arguments{
\item{x}{the output from tes() or tess()}
}
\value{
a plot
}
\description{
Plot fitted curve for TES and TESS
}
