% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TESS.R
\name{plot_tess}
\alias{plot_tess}
\title{Plot fitted curve for TESS}
\usage{
plot_tess(TESS_output, ...)
}
\arguments{
\item{TESS_output}{the output from tess()}

\item{...}{other arguments passed to plot()}
}
\value{
a plot
}
\description{
Plot fitted curve for TESS
}
