% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes.R
\name{print.rarestr}
\alias{print.rarestr}
\title{Print the "rarestr" class}
\usage{
\method{print}{rarestr}(x, ...)
}
\arguments{
\item{x}{a "rarestr" object#'}

\item{...}{Other arguments passed to print().}
}
\value{
Print the "rarestr" class
}
\description{
This function prints the contents of a rarestr object.
}
\examples{
data(share, package = 'rarestR')
Output_tes <- tes(share[1,])
Output_tes
}
