% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rarefaction}
\alias{rarefaction}
\title{Dataset for rarestR.}
\format{
A matrix:
\describe{
\item{columns}{A column is...}
\item{row}{A row is}
...
}
}
\usage{
rarefaction
}
\description{
A data set.
}
\keyword{datasets}
