% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S123}
\alias{S123}
\title{Dataset for rarestR.}
\format{
A matrix:
\describe{
\item{columns}{A column is...}
\item{row}{A row is}
...
}
}
\usage{
S123
}
\description{
A data set.
}
\keyword{datasets}
