% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESS.R
\name{tess}
\alias{tess}
\title{Calculate the Total number of Expected Shared Species between two samples.}
\usage{
tess(x, knots = 40, plot = FALSE)
}
\arguments{
\item{x}{a data matrix for two samples representing two communities (plot x species)}

\item{knots}{specifies the number of separate sample sizes of increasing value used for the calculation of ESS between 1 and the endpoint, which by default is set to knots=40}

\item{plot}{a logical value that provides the fitted curve}

\item{method}{the method the model used, with three options available as "auto", "logistic" and "weibull", with the default set as "auto"}
}
\value{
The asymptotic value (a), the model used in the estimation of TESS (either "Three" or "Four") and the model fit (r.sq).
}
\description{
Calculate the Total number of Expected Shared Species between two samples.
}
