% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES.R
\name{es}
\alias{es}
\title{Calculate the Expected Species}
\usage{
es(x, m, method = c("a", "b"), MARGIN = 1)
}
\arguments{
\item{x}{a data vector representing number of individuals for each species}

\item{m}{the sample size parameter that represents the number of individuals randomly drawn from the sample, which by default is set to m=1, but can be changed according to the users' requirements. For ESa, m can not be larger than the sample size}

\item{method}{the calculation approach of Expected Species used, with two options available as "a" and "b" to calculate ESa and ESb, with the default set as "a"}

\item{MARGIN}{}
}
\value{
a value of Expected Species
}
\description{
Calculate the Expected Species
}
\examples{
es(S123[1, ], m = 300)
es(S123[1, ], m = 2)
es(S123[1, ], m = 100)
es(S123[2, ], m = 100)
es(S123[2, ], m = 100, method = "b")
es(S123[2, ], m = 100, method = "a")

es(S123, m = 200, MARGIN = 1)
es(S123, m = 2, MARGIN = 2)

es(S123[, 1:3])
es(S123[1:3, ], m = 150)

es(S123[1:3, ], m = 20, "b")
}
