% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rarestr.R
\name{plot.rarestr}
\alias{plot.rarestr}
\title{Plot the "rarestr" class}
\usage{
\method{plot}{rarestr}(x, ...)
}
\arguments{
\item{x}{a "rarestr" object}

\item{...}{other arguments passed to plot()}
}
\value{
Plot the "rarestr" class
}
\description{
Plot the "rarestr" class
}
\examples{
data(share, package = 'rarestR')
Output_tes <- tes(share[1,])
Output_tes
plot(Output_tes)
}
