% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TES.R
\name{plot_tes}
\alias{plot_tes}
\title{Plot fitted curve for TES}
\usage{
plot_tes(TES_output, ...)
}
\arguments{
\item{TES_output}{the output from tes()}

\item{...}{other arguments passed to plot()}
}
\value{
a plot
}
\description{
Plot fitted curve for TES
}
